Bug 21444: Adds the permanent location
To Test:
1) Be sure ReturnToShelvingCart is set to Move.
2) Check out an item.
3) Check in the item.
4) Note that only the CART location shows in the Shelving location
column.
5) Apply the patch.
6) Check out an item.
7) Check in the item.
8) Note that the permant location now shows with the cart location next
to in in parenthesis.
9) Try combinations of missing descriptions or missing locations from
the LOC table for the item permanent location and cart.  If there is no
description, it should show the code in its place.  If there is no
location defined, it will show either the code that is stored in the
item or be blank.